deps/v8/test/inspector/debugger/suspended-generator-scopes-expected.txt

Tests that suspended generators produce scopes

Running test: testScopesPaused
{
    id : <messageId>
    result : {
        result : [
            [0] : {
                configurable : true
                enumerable : true
                isOwn : true
                name : a
                value : {
                    description : 420
                    type : number
                    value : 420
                }
                writable : true
            }
            [1] : {
                configurable : true
                enumerable : true
                isOwn : true
                name : b
                value : {
                    description : 42
                    type : number
                    value : 42
                }
                writable : true
            }
        ]
    }
}

Running test: testScopesNonPaused
{
    id : <messageId>
    result : {
        result : [
            [0] : {
                configurable : true
                enumerable : true
                isOwn : true
                name : a
                value : {
                    description : 430
                    type : number
                    value : 430
                }
                writable : true
            }
            [1] : {
                configurable : true
                enumerable : true
                isOwn : true
                name : b
                value : {
                    type : undefined
                }
                writable : true
            }
        ]
    }
}
